+++ src/test/java/com/callisto/CallistoApplicationTests.java
@@ -20,44 +20,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@SpringBootTest
 class CallistoApplicationTests {
-
-
-    @Autowired
-    TestRestTemplate restTemplate;
-
-    @MockitoBean
-    UserService userService;
-
-    @Test
-    void testRegisterUser() throws Exception {
-
-        Address address = new Address();
-        address.setStreet("hanuman nagar");
-        address.setCity("Patna");
-        address.setState("Bihar");
-        address.setCountry("India");
-        address.setPin("800020");
-
-        User user = new User();
-        user.setFirstName("S");
-        user.setMiddleName("");
-        user.setLastName("kumar");
-        user.setDob(LocalDate.of(1995, 9, 24));
-        user.setEmail("kumarsuraj15@gmail.com");
-        user.setPhoneNumber("7296098302");
-        user.setPassword("Password@123");
-        user.setAddress(address);
-        user.setTermsAccepted(true);
-
-
-        doNothing().when(userService).saveUser(any());
-        ResponseEntity<String> response = restTemplate.postForEntity("/api/user/register",user,String.class);
-        System.out.println(response);
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-        assertThat(response.getBody()).isEqualTo("User registered successfully");
-    }
-
 }
