+++ src/test/java/com/callisto/userServiceTest/UserRegisterServiceTest.java
@@ -1,5 +1,4 @@
-package com.callisto;
-
+package com.callisto.userServiceTest;
 
 import com.callisto.exception.FirstNameNotFoundException;
 import com.callisto.exception.UserEmailNotFoundException;
@@ -16,10 +15,13 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-public class UserServiceTest {
+public class UserRegisterServiceTest {
 
     @Mock
     private UserRepository userRepository;
@@ -82,4 +84,3 @@ public class UserServiceTest {
         verify(userRepository).save(validUser);
     }
 }
-
